+++ atmega_src/main.c
@@ -6,6 +6,9 @@
 #include "servo.h"
 #include "ir.h"
 #include "servo.h"
+#include "main.h" 
+#define E_FAIL (-1)
+#define S_OK (0)
 
 //Button on PA0 mocks authorization signal to unlock
 #define AUTHORIZED ~PINA & 0x01
@@ -146,6 +149,7 @@ int lock_Tick(int lock_state)
 
 int main(void)
 {
+	
 	//PA0 - "Authorize" Mock button, PA1-3 Package IR sensors, PA4 Door IR Sensor
 	DDRA = 0x00;
 	PORTA = 0xFF;
@@ -155,6 +159,11 @@ int main(void)
 	//Debug light on PC0
 	DDRC = 0xFF;
 	PORTC = 0x00;
+	
+	DDRD = 0xFF;
+	PORTD = 0x00;
+	
+	run_tests();
 
 	TimerSet(1);
 	TimerOn();
@@ -190,6 +199,8 @@ int main(void)
 	unsigned short i; // Scheduler for-loop iterator
 	while (1)
 	{
+		 
+		 
 		// Scheduler code
 		for (i = 0; i < numTasks; i++)
 		{
