+++ atmega_src/main.h
@@ -1,0 +1,16 @@
+/*
+ * main.h
+ *
+ * Created: 5/26/2020 10:11:32 PM
+ *  Author: Christian
+ */ 
+
+
+#ifndef MAIN_H_
+#define MAIN_H_
+
+void run_tests();
+
+
+
+#endif /* MAIN_H_ */
